+++ packages/syntax-core/src/Button/Button.tsx
@@ -12,6 +12,7 @@ import iconSize from "./constants/iconSize";
 import textVariant from "./constants/textVariant";
 import loadingIconSize from "./constants/loadingIconSize";
 import styles from "./Button.module.css";
+import useIsHydrated from "../useIsHydrated";
 
 type ButtonProps = {
   /**
@@ -39,7 +40,8 @@ type ButtonProps = {
     | "destructive-secondary"
     | "destructive-tertiary"
     | "branded"
-    | "success";
+    | "success"
+    | "inverse";
   /**
    * The size of the button
    *
@@ -117,6 +119,8 @@ const Button = forwardRef<HTMLButtonElement, ButtonProps>(
     }: ButtonProps,
     ref,
   ) => {
+    const isHydrated = useIsHydrated();
+
     return (
       <button
         data-testid={dataTestId}
@@ -124,7 +128,7 @@ const Button = forwardRef<HTMLButtonElement, ButtonProps>(
         aria-label={accessibilityLabel}
         type={type}
         title={tooltip}
-        disabled={disabled || loading}
+        disabled={!isHydrated || disabled || loading}
         onClick={onClick}
         className={classNames(
           styles.button,