+++ packages/syntax-core/src/TapArea/TapArea.stories.tsx
@@ -82,3 +82,60 @@ export const Rounding: StoryObj<typeof Box> = {
     </>
   ),
 };
+
+export const Colored: StoryObj<typeof Box> = {
+  render: () => (
+    <>
+      <Box display="flex" direction="column" gap={4}>
+        <Typography>Hover to see the overlay</Typography>
+
+        <TapArea
+          fullWidth={false}
+          onClick={() => {
+            /* empty */
+          }}
+        >
+          <Box
+            display="flex"
+            alignItems="center"
+            gap={2}
+            padding={2}
+            backgroundColor="purple300"
+          >
+            <Typography>Colored</Typography>
+          </Box>
+        </TapArea>
+      </Box>
+    </>
+  ),
+};
+
+export const NestedLink: StoryObj<typeof Box> = {
+  render: () => (
+    <>
+      <Box display="flex" direction="column" gap={4}>
+        <TapArea
+          fullWidth={false}
+          onClick={() => {
+            /* empty */
+          }}
+        >
+          <Box
+            display="flex"
+            alignItems="center"
+            gap={2}
+            padding={2}
+            backgroundColor="purple200"
+          >
+            <Typography>
+              Link:{" "}
+              <a href="https://www.cambly.com" target="_blank">
+                Cambly
+              </a>
+            </Typography>
+          </Box>
+        </TapArea>
+      </Box>
+    </>
+  ),
+};